Prevent loading security configuration twice
